+++ internal/args/args.go
@@ -6,8 +6,8 @@ import (
 	"fmt"
 	"reflect"
 
+	"github.com/cschleiden/go-workflows/backend/converter"
 	"github.com/cschleiden/go-workflows/backend/payload"
-	"github.com/cschleiden/go-workflows/converter"
 	"github.com/cschleiden/go-workflows/internal/sync"
 )
