+++ internal/workflowstate/workflowstate_test.go
@@ -5,8 +5,8 @@ import (
 	"testing"
 
 	"github.com/benbjohnson/clock"
+	"github.com/cschleiden/go-workflows/backend/converter"
 	"github.com/cschleiden/go-workflows/backend/payload"
-	"github.com/cschleiden/go-workflows/converter"
 	"github.com/cschleiden/go-workflows/core"
 	"github.com/cschleiden/go-workflows/internal/sync"
 	"github.com/google/uuid"
